feature(subcategory): bulk creation of sub category added

+++ Backend-Migration/cmd/migration/sql/000008_product_subcategories-table.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS product_subcategories (
   id BIGINT PRIMARY KEY AUTO_INCREMENT,
   product_id BIGINT NOT NULL,
-  subcategory_id INT NOT NULL,
+  subcategory_id BIGINT NOT NULL,
   FOREIGN KEY (product_id) REFERENCES products(id) ON DELETE CASCADE,
   FOREIGN KEY (subcategory_id) REFERENCES subcategories(id) ON DELETE CASCADE
 );
